Refactor export and download
